fix: substitute {{programName}} in ACK subject; dedupe ACK copy constants
Review fixes on the scholarship ack settings PR (thpr):

1. resolveAckCopy now substitutes {{programName}} in the subject, not
   just the body, mirroring the body's fallback-to-"" for a missing
   value. Subject is plain text, no HTML escaping. Documented the
   token in the subject field's settings-page description.

2. Extracted DEFAULT_ACK_SUBJECT/MEMBERSHIP_BODY/PROGRAM_BODY and
   renderAckBody into a new src/lib/scholarshipAckCopy.ts (depends
   only on the import-free, client-safe escapeHtml). scholarshipEmails.ts
   re-exports them so existing import paths keep working. The
   settings page now imports the real constants instead of
   maintaining a hand-duplicated copy that could drift.
