+++ checkin-app/src/lib/scholarshipEmails.ts
@@ -4,7 +4,8 @@ import type { HouseholdRecipient as ScholarshipRecipient } from "@/lib/emailReci
 import { parseEmailHeaderList } from "@/lib/emailHeader";
 import { logger } from "@/lib/logger";
 import prisma from "@/lib/prisma";
-import { escapeHtml } from "@/lib/email-templates/base";
+import { DEFAULT_ACK_SUBJECT, DEFAULT_ACK_MEMBERSHIP_BODY, DEFAULT_ACK_PROGRAM_BODY, renderAckBody } from "@/lib/scholarshipAckCopy";
+export { DEFAULT_ACK_SUBJECT, DEFAULT_ACK_MEMBERSHIP_BODY, DEFAULT_ACK_PROGRAM_BODY, renderAckBody };
 
 /**
  * Scholarship / payment-plan notification helpers. Small, dependency-light
@@ -19,33 +20,6 @@ import { escapeHtml } from "@/lib/email-templates/base";
  * Scholarship Review Team, not by an automated status email — see
  * docs/PROGRAM_CAPACITY_AND_SCHOLARSHIPS.md §5.
  */
-
-export const DEFAULT_ACK_SUBJECT = "We received your scholarship / payment-plan request";
-export const DEFAULT_ACK_MEMBERSHIP_BODY =
-    "Hi — we've received your household's scholarship / payment-plan request for your Treehouse membership dues. "
-    + "The Scholarship Review Team will review it and follow up.";
-export const DEFAULT_ACK_PROGRAM_BODY =
-    "Hi — we've received your scholarship / payment-plan request for {{programName}}. "
-    + "The Scholarship Review Team will review it and follow up. Your spot is held while they do.";
-
-/**
- * Render a plain-text ACK template into email-safe HTML: escape first (so a
- * `{{programName}}` value — or the template itself — can never inject markup),
- * substitute `{{programName}}` with its escaped value, then split on line
- * breaks into `<p>` paragraphs (blank lines dropped). Deliberately NOT an
- * HTML-authoring surface — no dangerouslySetInnerHTML, no preview iframe (see
- * the outreach template's sandboxed-iframe preview, settings/outreach/page.tsx,
- * for the XSS history this avoids by construction).
- */
-export function renderAckBody(template: string, vars: { programName?: string } = {}): string {
-    const escaped = escapeHtml(template).replaceAll("{{programName}}", escapeHtml(vars.programName ?? ""));
-    return escaped
-        .split(/\r\n|\r|\n/)
-        .map((line) => line.trim())
-        .filter((line) => line.length > 0)
-        .map((line) => `<p>${line}</p>`)
-        .join("");
-}
 
 type AckSettings = {
     scholarshipAckSubject?: string | null;
@@ -66,7 +40,8 @@ export function resolveAckCopy(
     vars: { programName?: string } = {},
 ): { subject: string; body: string } {
     const subjectRaw = settings?.scholarshipAckSubject?.trim();
-    const subject = subjectRaw || DEFAULT_ACK_SUBJECT;
+    // Mirrors renderAckBody's {{programName}} fallback (?? ""); subject is plain text, no HTML escaping.
+    const subject = (subjectRaw || DEFAULT_ACK_SUBJECT).replaceAll("{{programName}}", vars.programName ?? "");
 
     const bodyTemplateRaw = variant === "membership" ? settings?.scholarshipAckMembershipBody : settings?.scholarshipAckProgramBody;
     const bodyTemplate = bodyTemplateRaw?.trim()
